migrate from terraform-plugin-sdk/v2 to terraform-plugin-framework
The changes in this PR address the feedback given in
https://github.com/lifeomic/terraform-provider-phc/pull/9 (which was
mostly around the user not needing to put their comparisons in different
block types based on the attributes they use).

Because of the limitations I mentioned in that previous PR, these
changes weren't possible in the plugin-sdk. The plugin-framework is
supposed to solve a lot of the issues with inflexible schemas in the
previous sdk. It's *very* new -- only 3 real* upstream providers (
https://github.com/hashicorp/terraform-provider-salesforce,
https://github.com/hashicorp/terraform-provider-tls,
https://github.com/hashicorp/terraform-provider-random) maintained
by hashicorp use this. It's completely backwards compatible from the
users perspective.

The exact config that Swain sugguested works with this code. https://github.com/lifeomic/terraform-provider-phc/pull/9#pullrequestreview-1056003562

When this PR is merged I'll throw a follow up with the e2e tests.